Move signatures to own file
