+++ model/src/lib.rs
@@ -4,159 +4,16 @@ pub mod appdata_hexadecimal;
 pub mod h160_hexadecimal;
 pub mod order;
 pub mod ratio_as_decimal;
+pub mod signature;
 pub mod trade;
 pub mod u256_decimal;
 
 use ethabi::{encode, Token};
 use hex::{FromHex, FromHexError};
 use lazy_static::lazy_static;
-use primitive_types::{H160, H256};
-use serde::{de, Deserialize, Serialize};
+use primitive_types::H160;
 use std::fmt;
-use web3::{
-    signing::{self, Key, SecretKeyRef},
-    types::Recovery,
-};
-
-#[derive(Eq, PartialEq, Clone, Copy, Debug, Deserialize, Serialize, Hash)]
-#[serde(rename_all = "lowercase")]
-pub enum SigningScheme {
-    Eip712,
-    EthSign,
-}
-
-#[derive(Eq, PartialEq, Clone, Copy, Debug, Default, Hash)]
-pub struct Signature {
-    pub r: H256,
-    pub s: H256,
-    pub v: u8,
-}
-
-fn hashed_eip712_message(domain_separator: &DomainSeparator, struct_hash: &[u8; 32]) -> [u8; 32] {
-    let mut message = [0u8; 66];
-    message[0..2].copy_from_slice(&[0x19, 0x01]);
-    message[2..34].copy_from_slice(&domain_separator.0);
-    message[34..66].copy_from_slice(struct_hash);
-    signing::keccak256(&message)
-}
-
-fn hashed_ethsign_message(domain_separator: &DomainSeparator, struct_hash: &[u8; 32]) -> [u8; 32] {
-    let mut message = [0u8; 60];
-    message[..28].copy_from_slice(b"\x19Ethereum Signed Message:\n32");
-    message[28..].copy_from_slice(&hashed_eip712_message(domain_separator, struct_hash));
-    signing::keccak256(&message)
-}
-
-fn hashed_signing_message(
-    signing_scheme: SigningScheme,
-    domain_separator: &DomainSeparator,
-    struct_hash: &[u8; 32],
-) -> [u8; 32] {
-    match signing_scheme {
-        SigningScheme::Eip712 => hashed_eip712_message(domain_separator, struct_hash),
-        SigningScheme::EthSign => hashed_ethsign_message(domain_separator, struct_hash),
-    }
-}
-
-impl Signature {
-    /// r + s + v
-    pub fn to_bytes(self) -> [u8; 65] {
-        let mut bytes = [0u8; 65];
-        bytes[..32].copy_from_slice(self.r.as_bytes());
-        bytes[32..64].copy_from_slice(self.s.as_bytes());
-        bytes[64] = self.v;
-        bytes
-    }
-
-    pub fn from_bytes(bytes: &[u8; 65]) -> Self {
-        Signature {
-            r: H256::from_slice(&bytes[..32]),
-            s: H256::from_slice(&bytes[32..64]),
-            v: bytes[64],
-        }
-    }
-
-    pub fn validate(
-        &self,
-        signing_scheme: SigningScheme,
-        domain_separator: &DomainSeparator,
-        struct_hash: &[u8; 32],
-    ) -> Option<H160> {
-        let message = hashed_signing_message(signing_scheme, domain_separator, struct_hash);
-        let recovery = Recovery::new(message, self.v as u64, self.r, self.s);
-        let (signature, recovery_id) = recovery.as_signature()?;
-        signing::recover(&message, &signature, recovery_id).ok()
-    }
-
-    pub fn sign(
-        signing_scheme: SigningScheme,
-        domain_separator: &DomainSeparator,
-        struct_hash: &[u8; 32],
-        key: SecretKeyRef,
-    ) -> Self {
-        let message = hashed_signing_message(signing_scheme, domain_separator, struct_hash);
-        // Unwrap because the only error is for invalid messages which we don't create.
-        let signature = key.sign(&message, None).unwrap();
-        Self {
-            v: signature.v as u8,
-            r: signature.r,
-            s: signature.s,
-        }
-    }
-}
-
-impl Serialize for Signature {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        let mut bytes = [0u8; 2 + 65 * 2];
-        bytes[..2].copy_from_slice(b"0x");
-        // Can only fail if the buffer size does not match but we know it is correct.
-        hex::encode_to_slice(&self.to_bytes(), &mut bytes[2..]).unwrap();
-        // Hex encoding is always valid utf8.
-        let str = std::str::from_utf8(&bytes).unwrap();
-        serializer.serialize_str(str)
-    }
-}
-
-impl<'de> Deserialize<'de> for Signature {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        struct Visitor {}
-        impl<'de> de::Visitor<'de> for Visitor {
-            type Value = Signature;
-
-            fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-                write!(formatter, "the 65 signature bytes as a hex encoded string")
-            }
-
-            fn visit_str<E>(self, s: &str) -> Result<Self::Value, E>
-            where
-                E: de::Error,
-            {
-                let s = s.strip_prefix("0x").ok_or_else(|| {
-                    de::Error::custom(format!(
-                        "{:?} can't be decoded as hex signature because it does not start with '0x'",
-                        s
-                    ))
-                })?;
-                let mut bytes = [0u8; 65];
-                hex::decode_to_slice(s, &mut bytes).map_err(|err| {
-                    de::Error::custom(format!(
-                        "failed to decode {:?} as hex signature: {}",
-                        s, err
-                    ))
-                })?;
-                Ok(Signature::from_bytes(&bytes))
-            }
-        }
-
-        deserializer.deserialize_str(Visitor {})
-    }
-}
+use web3::signing;
 
 /// Erc20 token pair specified by two contract addresses.
 #[derive(Eq, PartialEq, Copy, Clone, Debug, Hash, Ord, PartialOrd)]
